Update: Reverted moving all into index.php
The loading times were long as the server had to check multiple user
types and then display the correct information.

Easier for group members to break tasks down my working on a page using
similar styling. Doesnt select data that isnt required. Returns user to
the correct page if they try to url jump.

The admin is currently a super admin and can also edit comments. This is
so code modules can be copied throughout the reader and author account.
Once the admin-index is finished then can carry over code to the correct
user page.